privora-api/requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
python-dotenv
passlib
cryptography